Working PreconditionViolation, but no extra properties yet.
This needs the "culprit" (calling) function as extra parameter. ConditionError needs the subject function as extra parameter (to find where it went wrong). This is difficult. The contractFunction itself doesn't make much sense. It is the same everywhere. The implFunction is a better candidate. It is also the culprit for post and exception. But, if the contractFunction has a name or a nom, we should send that too.

To use that, we need separate verifyOne methods for pre, post, exception. Post has as extra the result, exception the exception. The "extra arguments" should go there.
It probably makes little sense to keep verifyAll now.