Contract functions are now also required to have a location.
Made tests work by adding a location. We are missing tests for when we don’t have a location, or the location is not valid.
The contract dummy implementation now has the location.
ImplementableContract.implementation now adds the location to the contract function.

TODO: tests for when the location is not correct, both for Contract and contract function.