Experiment in using jsdoc. Run `npm run doc`. See jsdoc.json for configuration.
Need to add a lot of @… stuff, and the way modules are defined is counter productive, and should be brought in line, so that can also be used with amd.
Stuff now appears double, and is not clean.

Con is needed to avoid excluding _ directories. But, that is just for this test. Actually, we do not want to expose the _private stuff.

Then, documentation should be cleaned up. Sadly, methods that have no /** */ don’t show up at all. So, if we have nothing to say, we still need to say something :-(.