Commit 5130a50575dc2867efe95f45b7aa005bcf57dbcb introduced a test error in Firefox 52 and earlier. This will be fixed in FF 53. Documented.
Anonymous functions get the name of the variable they are declared in in ES2015. This is, at 2017-03-26, implemented in node, Chrome and Safari on Mac, but not yet in Firefox 52.
Firefox 53, based on Gecko 53, will ship in April 2017, and will have this functionality: https://developer.mozilla.org/en-US/Firefox/Releases/53