Add a "test" to show a CommonJs load error with the current library
It turned out that, when loading AbstractContract (indirectly via Contract), the library failed when creating
AbstractContract.root, because there was no "location outside the library" from which it was called
(defined as not in ../src, and not native code).

This cannot be tested with a test framework, because that does include non-native code in the stack.
So this "test" is a program to be called in addition to the unit tests.

This now fails.