Add more stringent conditions on what it means to be a contract function, to support constructors
Now tests don't work, or will fail. This means that to support constructors, we need to change several implementations.

Also, since the condition to be a contract function is more stringent, other implementations outside this library that would implement that requirement, will no longer work. This means this change is incompatible, although our code will adhere to the new conditions in the next version. To keep backward compatibility, we might add the new condition separately. It could then be added, or strengthened, in postconditions. However, we would then not be able to build on it in preconditions. Furthermore, the separation has no meaningful semantics in the resulting code. We will therefor make an incompatible change, and change the mark version of the library, although for most existing uses of the library the change will be compatible.