Bump mark version from II to III
This all tests GREEN, except Safari, that keeps complaining about a load timeout on just.randomstring.

In semantic versioning, this will be 3.0.0 (earlier, we started at 1.0.0 with a new mark version, and disregarded the micro version, but that is not ok). But the package name changes too. Still looking what is the best way to do this with npm.