Change is not a promise determination to use verifyAll like the other determinations
fails when it does not return a Promise now is GREEN - the reason was the location

Although we can test for the promise being a Promise directly, the current implementation chooses to go through verifyAll. This is to keep the stack determination in 1 place. If we would throw the PostconditionViolation here, we need to define the stack trace here for all violations, and pass it through verifyAll and verify. For this to work, we need to create an applicable extendedArgs (without tainting the original for uses below). This is at the cost of going through verifyAll and verify, and the resultIsAPromiseCondition to do a simple test.