This demonstrates an issue that only occurs in Safari on iOS
Execute
		> npm run wd
to see

User Agent: "Mozilla/5.0 (iPhone; CPU iPhone OS 11_2 like Mac OS X) AppleWebKit/604.4.7 (KHTML, like Gecko) Version/11.0 Mobile/15C114 Safari/604.1"

The issue occurs in AbstractContract.isAGeneralContractFunction. Console.log shows that the prototype of the contract function is not as expected an instance of the prototype.

The weird thing is that this does not fail when adding a console.log line in test/02.IV/10.1.PromiseContractFunctionTest.js fibonacciImpl (now commented out), or when adding an assert to lib/IV/AbstractContract.js bless (now commented out).

Since this only occurs in Safari on iOS, it is considered a bug in that browser. I currently have no better explanation.