change test that determines whether or not we set the contract function prototype for promise contract functions
in Safari on iOS, tests show that 'most of the time' this prototype is not set in our tests, as it should be. It seems to depend on the complexity of the function. If a log command is added, the prototype is set. To work around this, this test is replaced with !!implFunction.prototype. This defaults to the prototype set in Function.prototype, which is an Object. This means we now replace the contractFunction prototype more often than needed, but that is not a functional problem.