Fix returning Safari-on-iOS problem
The problem of 4ed9879c6b5544b174ae0825d7f7055fd5e147d8 was back.
This time, traced it down to the bless test.

implFunction DOES NOT have a prototype (tested). By doing a must on it, IT HAS. It seems that the prototype of a function is added lazily on this platform ?!?!?!. Adding this test makes sure it is there (it forces it, and tests it).