try to type the public interfaces of contracts-v, and compile a first test
over 2 days of work, still not there

the types here are quite intricate, and simplification has been tried
of course it can be more simple, what then, what is the sense of doing it?

TS takes for more time than it helps when coding
once this is done, it will help with writing contracts in an IDE in JS
that is the intention

what is annoying, is that with the very strict compiler settings, the
existing JS code is validated (good) but fails too much

`"noImplicitAny": false` helps, for the existing JS code
now the TS compiler requires types on JS code
either we diversify tsconfig settings for the d.ts file and the d.ts tests,
or we add types with JSDoc. Is that possible? On the other hand, why
doesn't it use the index.d.ts specifications?