try to use js inheritance in ts
Let tsc generate t.ds for jsSubjects with `> npm run tsc`

In the generated `.t.ds` we see that TS detects the "classes" Super- and SubConstructor, but not the inheritance relationship, from JS.

The inheritance relationship is detected by TS from JS when we use the `class` syntax.

Note also that the generated code `.js` for the SubConstructor is invalid JS: the `constructor` property we set in the prototype, as has to be done in JS, is added as a property to the generated `class` based JS, but there it is a reserved word!

(`npm run tsd` to see this)

In other words, to be able to use "JS inheritance" in TS, we _need_ to create a t.ds that makes the inheritance of constructor functions explicit. If we generate it, we have to edit it to make the inheritance of constructor functions explicit.