rm jsSubjects
The previous comment shows that we only can define inheritance with "class" definitions, in JS, or TS.

Therefor, it is impossible to test whether SubPrototype would work correctly with JS constructor functions, because we cannot define TS inheritance with JS constructor functions. We need TS always for that.

Therefor, it makes no sense to defined JS subjects for testing SubPrototype.