Create a correct (?) generic definition of bind …
but call it bind2 in Callable for the moment

This adds to the overload of bind in CallableFunction, and does not override the existing definitions.

Tried to Omit<…> the CallableFunction#bind definitions, but Omit only copies the properties, and we loose the function signature with Omit.

Tried to build the signature again and add it, but that also adds bind again.