fix a test
Why did this start to fail? because we changed the name of the file today!
It never ran before!

When this was first coded, it probably worked with an older version of
`node:util`, and the output of `inspect` has since changed? That older version
(18? 10?, 12?) now is no longer supported by this code base?

[skip ci]

there are still tests failing